move single eds to testing file
